pyproject.toml
[project]
name = "pysatimage"
version = "0.0.16"
authors = [{ name = "Artur Shevtsov", email = "artorius.ua@gmail.com" }]
description = "SID: Satellite Imagery Downloader"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.32.3",
    "numpy>=2.0.0",
    "opencv-python>=4.8.0",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.2",
    "setuptools>=75.7.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/artoriusss/sid"
Issues = "https://github.com/artoriusss/sid/issues"
